requirements.txt
fastapi[standard]>=0.115.12
uvicorn>=0.31.0
sqlalchemy>=2.0.40
pytest>=8.3.5
alembic>=1.15.2
aiohttp>=3.11.16
python-dotenv~=1.0.1
aiosqlite~=0.21.0
pre-commit~=4.2.0
